+++ README.md
@@ -1,21 +1,24 @@
 # 조사해보기
-## Client Side Rendering이란? 전통적인 SSR 방식의 웹앱과의 차이점은?
 
-CSR(Client Side Rendering): Renderring에 필요한 모든 정보를 받아오지 않고, 필요한 정보만 요청 후 변화된 부분만 rendering한다.  
-사용자의 행동에 따라 필요한 부분만 다시 읽어들이기 때문에 SSR보다 조금 더 빠른 상호작용이 가능하다.  
-page전체가 아닌 필요한 부분만 변경하여 모바일에서도 빠른 속도로 렌더링 가능하다.  
-다만, HTML다운로드 후 JS, CSS, API 등을 읽고 화면을 그리는 시간까지 모두 마쳐야 콘텐츠가 사용자에게 보이기 때문에 초기구동 속도가 느리다. 
+## 컴포넌트를 설계하고 나누는 기준에는 어떤 것들이 있을까요?
+재사용성, 가독성, 단일 책임 원칙(하나의 명확한 역할을 수행), 테스트 하기 쉬운 컴포넌트, 유지보수성 ...
 
-## 선언적 프로그래밍과 명령형 프로그래밍의 차이
-- 선언적 프로그래밍: 프로그램의 제어 흐름을 명시적으로 지정하지 않음(목적 지향, 절차 생략)
-- 명령형 프로그래밍: 프로그램이 어떻게 해야하는지 명령을 통해 단계별로 명시하여 상태를 변경함
+## 클래스 컴포넌트를 지양하게 된 이유에는 어떤 것들이 있을까요?
+코드의 복잡성과 가독성 저하(상속 구조, 복잡한 생명주기 메서드드), 그리고 함수형 컴포넌트와 비교했을 때, 개발 편의성 및 재사용성 저하 때문이다. 함수형 컴포넌트는 간단하고 명확한 문법을 가지며 훅을 통해 상태관리와 생명 주기 기능을 활용할 수 있기 때문에 React에선 클래스형 컴포넌트보다 함수형 컴포넌트와 훅을 주로 사용하는 것을 권장한다. 
 
-선언적 프로그래밍은 명령형 프로그래밍에 비해 코드가 직관적이고 읽기 쉽다. 각 요소가 무엇을 의미하는지 명확하게 드러낸다. 결과물에만 집중하고, 복잡한 과정은 추상화하여 신경쓰지 않는다.  
+## 클래스 컴포넌트를 반드시 사용해야 하는 경우는 언제일까요?
+오래된 라이브러리가 클래스 컴포넌트를 요구할 때.
 
-## 브라우저가 웹 페이지를 표시하는 과정
-URL 처리(html이 실제로 저장되어 있는 서버에 접근하기 위해) -> DNS 조회 (도메인 이름을 IP주소로 변경) -> 요청 전송(Https or Http) -> 응답 전송 -> 브라우저가 서버로부터 받은 자료를 바탕으로 렌더링
+## 컴포넌트 간에는 부모-자식 관계가 존재할 수 있습니다. 어떨 때 부모-자식 관계가 성립하는 걸까요?
+부모 컴포넌트가 데이터를 관리하고 자식 컴포넌트에게 전달하는 역할을 할 때, 
+부모 컴포넌트에서 import로 자식 컴포넌트를 불러올 때.
 
-## 리액트를 사용하는 이유
-- virtual dom: 실제 dom의 복제본인 virtual dom을 사용한다. 변화가 필요한 부분만 virtual dom에서 수정 후 실제 dom과 차이점을 찾아서 그 부분만 실제 dom에 반영한다. 
-- event handler: 브라우저가 처리해야 할 특정 사건이 발생하면 이를 감지하여 이벤트를 발생시킨다. 이때 호출되는 함수를 이벤트 핸들러라고 한다. 리액트에선 dom조작 없이 상태 기반으로 이벤트 처리를 할 수 있다.
-- component: 각 컴포넌트는 ui, 상태, 이벤트 핸들러를 독립적으로 관리할 수 있다. 
+## 배열을 렌더링할 때, 각 컴포넌트에 key값을 만들어 전달해줘야 합니다. key는 왜 필요한 걸까요? key값을 정하는 기준은 무엇인가요?
+key가 있어야 배열이 업데이트 될 때, 가리키는 객체가 명확해져서 효율적으로 렌더링할 수 있다. 
+만약 key설정을 하지 않게된다면 중간의 값이 바뀌었을 때, 하윗값들이 전부 변하게된다. 
+
+key는 배열의 각 원소들이 가지고 있는 고유값으로 설정한다. 
+
+## 조건부 렌더링이란 무엇일까요? 어떨 때 사용할 수 있을까요? 
+조건부 렌더링이란 특정 조건에 따라 다른 결과물을 렌더링하는 것이다. 즉, 원하는 동작을 캡슐화하는 컴포넌트를 만들 수 있다.
+상황에 따라 다른 표시를 해야하거나 특정 이벤트가 발생했을 때, 특정 요소를 표시하는 경우 등 이때 사용할 수 있다.
